Updates from snap 2.7.1

+++ src/snaplib/dbl4/snap_dbl4_interface.c
@@ -20,7 +20,7 @@ static int lastpercent = 0;
 static char laststate[MAX_MSG_LEN+1] = {0};
 static int tracing = -1;
 
-void * utlAlloc( int size)
+void * utlAlloc( size_t size)
 {
     return check_malloc( size );
 }
@@ -143,7 +143,7 @@ StatusType utlSeekBlobDB( void *blob, long lngOffset, int whence )
 
 StatusType utlTellBlobDB( void *blob, long *lngOffset )
 {
-    (*lngOffset) = ftell( (FILE *) blob );
+    (*lngOffset) = (long) ftell( (FILE *) blob );
     return STS_OK;
 }
